+++ TractorMod/Framework/Attachments/AxeAttachment.cs
@@ -92,17 +92,17 @@ namespace Pathoschild.Stardew.TractorMod.Framework.Attachments
             // cut resource stumps
             if (this.Config.ClearDebris || this.Config.CutGiantCrops)
             {
-                ResourceClump clump = this.GetResourceClumpCoveringTile(location, tile);
+                ResourceClump clump = this.GetResourceClumpCoveringTile(location, tile, player, out var applyTool);
 
                 // giant crops
                 if (this.Config.CutGiantCrops && clump is GiantCrop)
-                    this.UseToolOnTile(tool, tile, player, location);
+                    applyTool(tool);
 
                 // big stumps and fallen logs
                 // This needs to check if the axe upgrade level is high enough first, to avoid spamming
                 // 'need to upgrade your tool' messages. Based on ResourceClump.performToolAction.
                 if (this.Config.ClearDebris && clump != null && this.ResourceUpgradeLevelsNeeded.ContainsKey(clump.parentSheetIndex.Value) && tool.UpgradeLevel >= this.ResourceUpgradeLevelsNeeded[clump.parentSheetIndex.Value])
-                    this.UseToolOnTile(tool, tile, player, location);
+                    applyTool(tool);
             }
 
             // cut bushes in large terrain features