+++ TractorMod/release-notes.md
@@ -1,5 +1,7 @@
 # Release notes
-## Upcoming release
+## 4.11
+Released 04 July 2020.
+
 * Added support for Generic Mod Config Menu (thanks to collaboration with NexusFlight!).
 * Added compatibility with boulders and stumps added by Farm Type Manager.
 * Fixed warp loop when riding the tractor onto the Witch's Hut warp runes.